tcConsoleBackend/pyproject.toml
[project]
name = "binfk-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.12.0",
    "bitsandbytes>=0.49.1",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.128.0",
    "fastapi-mail>=1.6.1",
    "ipykernel>=7.1.0",
    "pandas>=3.0.0",
    "protobuf>=6.33.4",
    "pymongo>=4.16.0",
    "pypdf>=6.7.0",
    "pypdf2>=3.0.1",
    "python-multipart>=0.0.21",
    "sentence-transformers>=5.2.3",
    "sentencepiece>=0.2.1",
    "sqlalchemy>=2.0.45",
    "sqlmodel>=0.0.31",
    "torch>=2.9.1",
    "transformers>=4.57.5",
    "uvicorn>=0.40.0",
    "weasyprint>=68.1",
]
